Diabetic/Diabetes/requirements.txt
flask
gunicorn
Cassandra-driver
pandas
numpy
matplotlib
seaborn
sklearn
scipy
cassandra-driver
apscheduler
datetime
nltk
